backend/requirements.txt
Pillow>=9.0.0
img2pdf>=0.5.0
python-docx>=0.8.11
fpdf2>=2.5.0
pdf2docx>=0.5.0
pypdf>=3.0.0
reportlab>=4.0.0
